requirements.txt
numpy
pandas
sklearn
torch
tick
faiss-cpu
faiss-gpu